Upgraded Xerces from v3.0.0 to v3.0.1 and added 64-bit version.
Updraded WinDbg files from v6.10.3.233 to v6.11.1.404 and added 64-bit version.

Removed Visual Studio 2008 Express solution and project files.

Implemented build scripts by simple bat files.

Ported the main application code to compile without warning on 64bit.

git-svn-id: file:///home/pauldoo/Programs/svnrepo@517 8e395407-42ff-0310-94d9-ec183df6106f

+++ Proffy/Proffy/WriteReport.cpp
@@ -68,7 +68,7 @@ namespace Proffy {
             for (std::set<PointInProgram>::const_iterator i = results->fEncounteredPoints.begin();
                 i != results->fEncounteredPoints.end();
                 ++i) {
-                const int id = std::distance(results->fEncounteredPoints.begin(), i);
+                const int id = static_cast<int>(std::distance(results->fEncounteredPoints.begin(), i));
                 
                 sourceFiles.insert(i->fFileName);
 
@@ -101,11 +101,11 @@ namespace Proffy {
                     const PointInProgram* const caller = i->first.first;
                     const PointInProgram* const callee = i->first.second;
                     const int callerId =
-                        std::distance(results->fEncounteredPoints.begin(),
-                            results->fEncounteredPoints.find(*caller));
+                        static_cast<int>(std::distance(results->fEncounteredPoints.begin(),
+                            results->fEncounteredPoints.find(*caller)));
                     const int calleeId = (callee == NULL) ? -1 :
-                        std::distance(results->fEncounteredPoints.begin(),
-                            results->fEncounteredPoints.find(*callee));
+                        static_cast<int>(std::distance(results->fEncounteredPoints.begin(),
+                            results->fEncounteredPoints.find(*callee)));
 
                     xercesc::DOMElement* const counter = document->createElement(L"Counter");
